refactor: maintenance components folder
