+++ apps/main/src/routes/RootLayout.tsx
@@ -15,7 +15,7 @@ import { TanStackRouterDevtools } from '@tanstack/react-router-devtools'
 import { CurveProvider } from '@ui-kit/features/connect-wallet'
 import { HydratorMap } from '@ui-kit/features/connect-wallet/lib/types'
 import { useWagmiConfig } from '@ui-kit/features/connect-wallet/lib/wagmi/useWagmiConfig'
-import { BackendMaintenanceModal } from '@ui-kit/features/maintenance/BackendMaintenanceModal'
+import { BackendMaintenanceModal } from '@ui-kit/features/maintenance/components/BackendMaintenanceModal'
 import { useMaintenance } from '@ui-kit/features/maintenance/hooks/useMaintenance'
 import type { Maintenance } from '@ui-kit/features/maintenance/hooks/useMaintenance'
 import { addBreadcrumb } from '@ui-kit/features/sentry'
