+++ packages/curve-ui-kit/src/features/maintenance/components/BackendMaintenanceBanner.tsx
