+++ packages/curve-ui-kit/src/features/maintenance/index.ts
@@ -1,4 +1,4 @@
-export * from './BackendMaintenanceBanner'
-export * from './BackendMaintenanceModal'
-export * from './MaintenancePage'
+export * from './components/BackendMaintenanceBanner'
+export * from './components/BackendMaintenanceModal'
+export * from './components/MaintenancePage'
 export * from './hooks/useMaintenance'
